Add basic debugging code through normal controller ports
USB geckos are impossible to obtain nowadays, and it's much
easier to plug an arduino into the controller port than into
the memory card slot, so that's how I've made my own debugging
setup. In the extremely unlikely case that you're reading this
commit message and are curious to know more, feel free to open
up an issue. -Wilm